# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stakeholders:
    get:
      tags: [Stakeholders]
      operationId: listStakeholders
      summary: List stakeholders
      responses:
        '200':
          description: All stakeholders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stakeholder'
        '401':
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Stakeholders]
      operationId: createStakeholder
      summary: Add a stakeholder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStakeholderRequest'
      responses:
        '200':
          description: The created stakeholder.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stakeholder'
  /stakeholders/{stakeholderId}:
    patch:
      tags: [Stakeholders]
      operationId: updateStakeholder
      summary: Update a stakeholder
      parameters:
        - name: stakeholderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStakeholderRequest'
      responses:
        '200':
          description: The updated stakeholder.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stakeholder'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Stakeholders]
      operationId: deleteStakeholder
      summary: Delete a stakeholder
      parameters:
        - name: stakeholderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir.
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Stakeholder:
      type: object
      required: [id, name, role, email, lastContacted, status, statusClass]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        email:
          type: string
        lastContacted:
          type: string
        status:
          type: string
        statusClass:
          type: string
    CreateStakeholderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
        email:
          type: string
        lastContacted:
          type: string
        status:
          type: string
        statusClass:
          type: string
    UpdateStakeholderRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
        email:
          type: string
        lastContacted:
          type: string
        status:
          type: string
        statusClass:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file.
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`).
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...